+++ app/src/main/kotlin/moe/pine/meteorshower/services/auth/AuthVerifyResult.kt
@@ -1,7 +1,7 @@
 package moe.pine.meteorshower.services.auth
 
 data class AuthVerifyResult(
-    val githubId: Int,
+    val githubId: Long,
     val name: String,
     val callbackUrl: String
 )
