+++ lib/auth/src/main/kotlin/moe/pine/meteorshower/auth/models/User.kt
@@ -4,7 +4,7 @@ import java.time.LocalDateTime
 
 data class User(
     var id: Long? = null,
-    var githubId: Int,
+    var githubId: Long,
     var name: String,
     var createdAt: LocalDateTime? = null,
     var updatedAt: LocalDateTime? = null
