+++ lib/github-auth/src/main/kotlin/moe/pine/meteorshower/github/auth/GitHubAuth.kt
@@ -97,7 +97,7 @@ class GitHubAuth {
         return GitHubAuthResult(
             accessToken = accessToken,
             userName = user.name,
-            userId = user.id
+            userId = user.id.toLong()
         )
     }
 }
